refactor: restructure database handling by moving to config module and integrating Redis support
