+++ Cargo.toml
@@ -5,7 +5,7 @@ edition = "2021"
 
 [dependencies]
 warp = "0.3"
-tokio = { version = "1", features = ["full"] }
+tokio = { version = "1.43", features = ["full"] }
 tokio-postgres = "0.7"
 serde = { version = "1", features = ["derive"] }
 serde_json = "1"
@@ -19,3 +19,4 @@ thiserror = "*"
 qrcode = "0.14.1"
 image = "0.25.5"
 base64 = "0.22.1"
+redis = {version = "*", features = ["tokio"]}
