+++ src/lib.rs
@@ -1,7 +1,7 @@
 pub mod handlers;
 pub mod models;
 pub mod routes;
-pub mod db;
+pub mod config;
 pub mod views;
 pub mod utils;
